Gujarat Election PowerBI Visual
Hi All, 
The above PDF file is a screenshot compilation of the PowerBI report created for the Gujarat Election of 2022. I have used Tooltip, ShapeMap, TopoJSON and Excel to generate the metadata needed for the file. This is just an essential compilation. I will be updating it further for more advanced and beautiful features.